iOS-5: Фикс коммов
