+++ CakesHub/CakesHub/Streams/ProductStream/ProductDetail/RatingReviewsScreen/Model/ProductReviewsModel+Mock.swift
@@ -1,5 +1,5 @@
 //
-//  ProductReviewsModelMockData.swift
+//  ProductReviewsModel+Mock.swift
 //  CakesHub
 //
 //  Created by Dmitriy Permyakov on 04.02.2024.
@@ -38,18 +38,13 @@ private extension ProductReviewsModel.CommentInfo {
         ProductReviewsModel.CommentInfo(
             userName: "Helene Moore \(number)",
             date: "June 5, 2019",
-            description: .commentText,
+            description: """
+            The dress is great! Very classy and comfortable. It fit perfectly! I'm 5'7" and 130 pounds. I am a 34B chest. This dress would be too long for those who are shorter but could be hemmed. I wouldn't recommend it for those big chested as I am smaller chested and it fit me perfectly. The underarms were not too wide and the dress was made well.
+            """,
             countFillStars: number % 6,
             feedbackCount: number
         )
     }
-}
-
-private extension String {
-
-    static let commentText = """
-    The dress is great! Very classy and comfortable. It fit perfectly! I'm 5'7" and 130 pounds. I am a 34B chest. This dress would be too long for those who are shorter but could be hemmed. I wouldn't recommend it for those big chested as I am smaller chested and it fit me perfectly. The underarms were not too wide and the dress was made well.
-    """
 }
 
 #endif
